Moving partition tracking code to partition/tracking
